Move MotionBlindsData and MotionBlindsConfigEntry to coordinator module
